cheri: Fix typecasts in fs/ext4
- Reference counts are actually integers stored in a pointer.
- 32-bit cast missed by automatic conversions
- Fake a pointer from an instruction pointer address in tracing.

+++ fs/ext4/mballoc.c
@@ -468,7 +468,7 @@ static inline void *mb_correct_addr_and_bit(int *bit, void *addr)
 	addr = (void *) ((uintptr_t) addr & ~7UL);
 #elif BITS_PER_LONG == 32
 	*bit += ((unsigned long) addr & 3UL) << 3;
-	addr = (void *) ((unsigned long) addr & ~3UL);
+	addr = (void *) ((uintptr_t) addr & ~3UL);
 #else
 #error "how many bits you are?!"
 #endif
